+++ tests/Neo.L2.Explore.UnitTests/Usings.cs
@@ -1,0 +1,4 @@
+global using Microsoft.VisualStudio.TestTools.UnitTesting;
+global using Neo.L2;
+global using Neo.L2.Sdk;
+global using Neo.L2.Explore;